Move `build update` to `update build-agent`.
