+++ cmd/update.go
@@ -1,6 +1,10 @@
 package cmd
 
 import (
+	"fmt"
+	"os/exec"
+	"regexp"
+
 	"github.com/CircleCI-Public/circleci-cli/version"
 	"github.com/pkg/errors"
 	"github.com/rhysd/go-github-selfupdate/selfupdate"
@@ -27,7 +31,53 @@ func newUpdateCommand() *cobra.Command {
 		RunE:  installUpdate,
 	})
 
+	update.AddCommand(&cobra.Command{
+		Use:   "build-agent",
+		Short: "Update the build agent to the latest version",
+		RunE:  updateBuildAgent,
+	})
+
 	return update
+}
+
+var picardRepo = "circleci/picard"
+
+func updateBuildAgent(cmd *cobra.Command, args []string) error {
+	latestSha256, err := findLatestPicardSha()
+
+	if err != nil {
+		return err
+	}
+
+	Logger.Infof("Latest build agent is version %s", latestSha256)
+
+	return nil
+}
+
+// Still depends on a function in cmd/build.go
+func findLatestPicardSha() (string, error) {
+	outputBytes, err := exec.Command("docker", "pull", picardRepo).CombinedOutput() // #nosec
+
+	if err != nil {
+		return "", errors.Wrap(err, "failed to pull latest docker image")
+	}
+
+	output := string(outputBytes)
+	sha256 := regexp.MustCompile("(?m)sha256.*$")
+	latest := sha256.FindString(output)
+
+	if latest == "" {
+		return "", fmt.Errorf("failed to parse sha256 from docker pull output")
+	}
+
+	// This function still lives in cmd/build.go
+	err = storeBuildAgentSha(latest)
+
+	if err != nil {
+		return "", err
+	}
+
+	return latest, nil
 }
 
 func checkForUpdates(cmd *cobra.Command, args []string) error {
